removed cyclic dependency
